// src/Contexts/Moas/Companies/application/interfaces/ICompanyCreatorRequest.ts
export interface ICompanyCreatorRequest {
  id: string;
  name: string;
  socialReason: string;
  email: string;
  prefixNumber: string;
  number: string;
  streetAddress1: string;
  streetAddress2: string;
  city: string;
  region: string;
  country: string;
  postalCode: string;
  lat: number;
  long: number;
}